requirements_net.txt
asgiref==3.5.2
Django==4.1.6
django-debug-toolbar==4.0.0
django-filter==23.5
djangorestframework==3.14.0
pillow==12.0.0
pytz==2023.3.post1
sqlparse==0.4.2
tzdata==2024.1